+++ CHANGES.rst
@@ -31,6 +31,7 @@ Fixed:
 * docutils image directive: crash when encountering a width/height containing a
   decimal point (#251 by Karel Frajtak)
 * docutils inline images don't support width, height and scale options
+* crash on using characters for page numbering (PDF backend)
 
 
 Release 0.5.2 (2021-02-24)